requirements.txt

flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
boto3==1.34.0
google-generativeai==0.3.2
